+++ src/ACFHelper.php
@@ -14,13 +14,13 @@ class ACFHelper {
 			acf_add_options_page(array(
 				'page_title' => __('Site Options', 'lathe'),
 				'menu_title' => __('Site Options', 'lathe'),
-				'menu_slug' => __('site-options', 'lathe'),
+				'menu_slug' => 'site-options',
 				'capability' => 'edit_posts',
 				'redirect' => false
 			));
 		});
 
-		add_action( 'acf/input/admin_head', function() {
+		add_action('acf/input/admin_head', function() {
 			add_meta_box(
 				'site-options-actions', 
 				__('Actions', 'lathe'), 
@@ -30,19 +30,6 @@ class ACFHelper {
 				"acf_options_page",
 				"side"
 			);
-		});
-
-		add_action('wp_ajax_clear_twig_cache', function() {
-			if (check_ajax_referer('clear_twig_cache')) {
-				$loader = new Loader();
-				$loader->clear_cache_twig();
-			}
-		});
-
-		add_action('wp_ajax_flush_rewrites', function() {
-			if (check_ajax_referer('flush_rewrites')) {
-				flush_rewrite_rules();
-			}
 		});
 	}
 }